Update DS_423_Tune_Assignment.ipynb
Removed import to sklearn from kerastuner.tuners. It was not a used import and invalid due to the outdate of kerastuner. We should look to replace this module.

gen_layer_nodes updated to increase and decrease dependent upon negative node incrementation.

Fixed create_model() function to include n_layers+1 in for loop